requirements.txt
streamlit
ultralytics
pillow
pandas
fpdf
google-generativeai
